Refactor
Create base class for commands
Move common functionality for command_callbacks,
events etc. into respective core packages
