+++ model_service/application/command_callbacks/core/command_callback.py
