+++ model_service/application/commands/core/command.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class Command(BaseModel):
+    """Base class for commands"""
