+++ model_service/application/commands/create_feed_forward_network.py
@@ -1,7 +1,6 @@
-from pydantic import BaseModel
-
+from model_service.application.commands.core.command import Command
 from model_service.application.schemas.feed_forward_network_schema import FeedForwardNetworkSchema
 
 
-class CreateFeedForwardNetwork(BaseModel):
+class CreateFeedForwardNetwork(Command):
     feed_forward_network: FeedForwardNetworkSchema
