+++ model_service/application/commands/create_table_dataset.py
@@ -1,7 +1,6 @@
-from pydantic import BaseModel
-
+from model_service.application.commands.core.command import Command
 from model_service.application.schemas.table_dataset_schema import TableDatasetSchema
 
 
-class CreateTableDataset(BaseModel):
+class CreateTableDataset(Command):
     table_dataset: TableDatasetSchema
