+++ model_service/application/commands/create_training_session.py
@@ -1,7 +1,6 @@
-from pydantic import BaseModel
-
+from model_service.application.commands.core.command import Command
 from model_service.application.schemas.training_session_schema import TrainingSessionSchema
 
 
-class CreateTrainingSession(BaseModel):
+class CreateTrainingSession(Command):
     training_session: TrainingSessionSchema
