+++ model_service/application/commands/start_ffn_training.py
@@ -1,5 +1,5 @@
-from pydantic import BaseModel
+from model_service.application.commands.core.command import Command
 
 
-class StartFfnTraining(BaseModel):
+class StartFfnTraining(Command):
     training_session_id: str
