+++ model_service/application/integration_event_publishers/default_exchange_publisher.py
@@ -4,7 +4,7 @@ from typing import Generic, TypeVar
 import attrs
 from pika.adapters.blocking_connection import BlockingChannel
 
-from model_service.application.integration_events.integration_event import IntegrationEvent
+from model_service.application.integration_events.core.integration_event import IntegrationEvent
 
 TIntegrationEvent = TypeVar("TIntegrationEvent", bound=IntegrationEvent)
