+++ model_service/application/integration_events/feed_forward_network_created.py
@@ -1,6 +1,6 @@
 import attrs
 
-from model_service.application.integration_events.integration_event import IntegrationEvent
+from model_service.application.integration_events.core.integration_event import IntegrationEvent
 
 
 @attrs.define
