+++ model_service/application/integration_events/training_session_created.py
@@ -1,6 +1,6 @@
 import attrs
 
-from model_service.application.integration_events.integration_event import IntegrationEvent
+from model_service.application.integration_events.core.integration_event import IntegrationEvent
 
 
 @attrs.define
