+++ model_service/application/services/create_feed_forward_network.py
@@ -1,6 +1,6 @@
 from model_service.application.commands.create_feed_forward_network import CreateFeedForwardNetwork
 from model_service.application.integration_events.feed_forward_network_created import FeedForwardNetworkCreated
-from model_service.application.integration_events.publishers import publish_integration_event
+from model_service.application.integration_events.core.publishers import publish_integration_event
 from model_service.application.schema_mappers.map_feed_forward_network_from_schema import (
     map_feed_forward_network_from_schema,
 )
