+++ model_service/application/services/create_table_dataset.py
@@ -1,5 +1,5 @@
 from model_service.application.commands.create_table_dataset import CreateTableDataset
-from model_service.application.integration_events.publishers import publish_integration_event
+from model_service.application.integration_events.core.publishers import publish_integration_event
 from model_service.application.integration_events.table_dataset_created import TableDatasetCreated
 from model_service.application.schema_mappers.map_table_dataset_from_schema import map_table_dataset_from_schema
 from model_service.dependencies.dependency_management.provide import Dependency, provide
